dataacquisition/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
et-xmlfile==1.1.0
greenlet==1.1.2
idna==3.3
install==1.3.5
isodate==0.6.1
lxml==4.8.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
progressbar2==4.0.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-utils==3.1.0
pytz==2021.3
rdflib==6.1.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SPARQLWrapper==1.8.5
SQLAlchemy==1.4.32
urllib3==1.26.8
wikipedia==1.4.0
